2D turb almost done

+++ src/global/global.h
@@ -346,7 +346,7 @@ using duration_t  = double;
 using simtime_t   = double;
 using timestep_t  = std::size_t;
 using ncells_t    = std::size_t;
-using npart_t     = unsigned long long int;
+using npart_t     = unsigned long int;
 
 // index/number
 using index_t = const std::size_t;
